refactor: Migrate Video engine to directory engine.
